Create PluginSystem namespace and folder for plugin-related interfaces
