+++ Orthodrome/Assets/Scripts/Plugin Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0bd7eab889b4c90418aa7fe7c075b8b3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
