+++ PluginTemplate/SimplePlugin.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using PluginSystem;
 
 namespace PluginTemplate
 {
